+++ src/events/messageCreate/handleMessagePosted.js
@@ -1,4 +1,4 @@
-const { MESSAGE_THRESHOLD } = require('../../../config/config.json')
+const { MESSAGE_THRESHOLD } = require('../../../config/config.json') 
 const generateMarkovOutput = require('../../utils/discord/generateMarkovOutput');
 const handleBotMentioned = require('../../utils/discord/handleBotMentioned');
 const messages = [];
@@ -6,8 +6,9 @@ const messageSet = new Set();
 let messageCounter = 0;
 let previousMessage;
 
-module.exports = async (client, message, classifier) => {
+module.exports = async (message, classifier) => {
     if (message.content === previousMessage) return;
+
     handleBotMentioned(message);
     if (message.author.bot) return;
     messageCounter++;