+++ src/utils/python/messageReply.py
@@ -34,7 +34,7 @@ def main() -> None:
     model_inputs = tokenizer([prompt], return_tensors="pt").to(model.device)
     generated_ids = model.generate(
     **model_inputs,
-    max_new_tokens=420,
+    max_new_tokens=520,
     do_sample=True)
     output_ids = generated_ids[0][len(model_inputs.input_ids[0]):].tolist() 
     try: